Huge refactor.  Jumping to double because it happens to be better supported for the Java 8 parallel stream ops.

+++ src/main/java/com/josephcatrambone/aij/Matrix.java
@@ -13,38 +13,38 @@ import java.util.function.UnaryOperator;
 public class Matrix implements Serializable {
 	public int rows;
 	public int columns;
-	public float[] data;
+	public double[] data;
 
 	public static Matrix identity(int rows) {
-		return new Matrix(rows, rows, (a,b) -> { if(a==b) {return 1.0f;} else {return 0.0f;} } );
+		return new Matrix(rows, rows, (a,b) -> { if(a==b) {return 1.0;} else {return 0.0;} } );
 	}
 
 	public static Matrix ones(int rows, int columns) {
-		return new Matrix(rows, columns, (a,b) -> 1.0f);
+		return new Matrix(rows, columns, (a,b) -> 1.0);
 	}
 
 	public Matrix() {
 		this.rows = 0;
 		this.columns = 0;
-		this.data = new float[0];
+		this.data = new double[0];
 	}
 
 	public Matrix(int rows, int columns) {
 		this.rows = rows;
 		this.columns = columns;
-		this.data = new float[rows*columns];
+		this.data = new double[rows*columns];
 	}
 
-	public Matrix(int rows, int columns, float[] data) {
+	public Matrix(int rows, int columns, double[] data) {
 		this.rows = rows;
 		this.columns = columns;
 		this.data = data;
 	}
 
-	public Matrix(int rows, int columns, BiFunction<Integer, Integer, Float> initFunction) {
+	public Matrix(int rows, int columns, BiFunction<Integer, Integer, Double> initFunction) {
 		this.rows = rows;
 		this.columns = columns;
-		this.data = new float[rows*columns];
+		this.data = new double[rows*columns];
 		for(int r=0; r < rows; r++) {
 			for(int c=0; c < columns; c++) {
 				data[c+r*columns] = initFunction.apply(r, c);
@@ -52,33 +52,34 @@ public class Matrix implements Serializable {
 		}
 	}
 
-	public float get(int r, int c) {
+	public double get(int r, int c) {
 		return this.data[c + r*columns];
 	}
 
-	public void set(int r, int c, float n) {
+	public void set(int r, int c, double n) {
 		this.data[c + r*columns] = n;
 	}
 
-	public void elementOp_i(UnaryOperator<Float> op) {
+	public void elementOp_i(UnaryOperator<Double> op) {
+		
 		for(int i=0; i < data.length; i++) {
 			data[i] = op.apply(data[i]);
 		}
 	}
 
-	public Matrix elementOp(UnaryOperator<Float> op) {
+	public Matrix elementOp(UnaryOperator<Double> op) {
 		Matrix m = new Matrix(this.rows, this.columns, Arrays.copyOf(this.data, rows*columns));
 		m.elementOp_i(op);
 		return m;
 	}
 
-	public void elementOp_i(Matrix other, BinaryOperator<Float> op) {
+	public void elementOp_i(Matrix other, BinaryOperator<Double> op) {
 		for(int i=0; i < data.length; i++) {
 			data[i] = op.apply(data[i], other.data[i]);
 		}
 	}
 
-	public Matrix elementOp(Matrix other, BinaryOperator<Float> op) {
+	public Matrix elementOp(Matrix other, BinaryOperator<Double> op) {
 		Matrix m = new Matrix(this.rows, this.columns, Arrays.copyOf(this.data, rows*columns));
 		m.elementOp_i(other, op);
 		return m;
@@ -88,7 +89,7 @@ public class Matrix implements Serializable {
 		Matrix result = new Matrix(this.rows, other.columns);
 		for(int i=0; i < rows; i++) {
 			for(int j=0; j < other.columns; j++) {
-				float accumulator = 0;
+				double accumulator = 0;
 				for(int k=0; k < this.columns; k++) {
 					// TODO: This must be made generic.
 					accumulator += this.get(i,k)*other.get(k,j);
@@ -128,15 +129,15 @@ public class Matrix implements Serializable {
 		}
 	}
 
-	public float[] getRow(int r) {
-		float[] rowData = new float[this.columns];
+	public double[] getRow(int r) {
+		double[] rowData = new double[this.columns];
 		for(int j=0; j < this.columns; j++) {
 			rowData[j] = this.get(r, j);
 		}
 		return rowData;
 	}
 
-	public void setRow(int r, float[] data) {
+	public void setRow(int r, double[] data) {
 		assert(data.length == this.columns);
 		for(int j=0; j < this.columns; j++) {
 			this.set(r, j, data[j]);
@@ -153,7 +154,7 @@ public class Matrix implements Serializable {
 		sb.append("]");
 
 		StringJoiner sj = new StringJoiner(",", "[", "");
-		for(float f : data) {
+		for(double f : data) {
 			sj.add(""+f);
 		}
 		sb.append(sj.toString());
@@ -170,9 +171,9 @@ public class Matrix implements Serializable {
 		String[] dims = tokens[1].split(",");
 		String[] data = tokens[2].split(",");
 		Matrix m = new Matrix(Integer.parseInt(dims[0]), Integer.parseInt(dims[1]));
-		m.data = new float[m.rows*m.columns];
+		m.data = new double[m.rows*m.columns];
 		for(int i=0; i < m.data.length; i++) {
-			m.data[i] = Float.parseFloat(data[i]);
+			m.data[i] = Double.parseDouble(data[i]);
 		}
 		return m;
 	}
