+++ src/main/java/com/josephcatrambone/aij/nodes/PowerNode.java
@@ -3,17 +3,17 @@ package com.josephcatrambone.aij.nodes;
 import com.josephcatrambone.aij.Matrix;
 
 public class PowerNode extends Node {
-	float exponent;
+	double exponent;
 
 	public PowerNode() { super(); }
 
-	public PowerNode(Node input, float e) {
+	public PowerNode(Node input, double e) {
 		super(input);
 		this.exponent = e;
 	}
 
 	public Matrix forward(Matrix[] args) {
-		return args[0].elementOp(a -> (float)Math.pow(a, exponent));
+		return args[0].elementOp(a -> (double)Math.pow(a, exponent));
 	}
 
 	// z := elem(x, op)
@@ -22,7 +22,7 @@ public class PowerNode extends Node {
 		return new Matrix[]{
 			adjoint.elementOp(
 				forward[0].elementOp(
-					a -> exponent*(float)Math.pow(a, exponent-1.0f)
+					a -> exponent*Math.pow(a, exponent-1.0f)
 				),
 				(adj,x)->adj*x
 			)
@@ -32,6 +32,6 @@ public class PowerNode extends Node {
 	// Used to augment serialization.
 	public String extraDataToString() { return ""+exponent; };
 	public void extraDataFromString(String s) {
-		exponent = Float.parseFloat(s);
+		exponent = Double.parseDouble(s);
 	}
 }
